Move Publsh component, replace StyledLinks
