+++ components/clientComponents/form-builder/app/index.ts
@@ -5,7 +5,7 @@ export { PreviewNavigation } from "./navigation/PreviewNavigation";
 export { EditNavigation } from "./navigation/EditNavigation";
 export { Start } from "./Start";
 export { Preview } from "./Preview";
-export { Publish } from "./Publish";
+export { Publish } from "../../../../app/(gcforms)/[locale]/(form administration)/form-builder/[id]/publish/Publish";
 export { Published } from "./Published";
 export { Edit } from "./edit";
 export { ShareModal } from "./ShareModal";
